+++ Modules/core/Controller/CorespaceController.php
@@ -30,7 +30,6 @@ class CorespaceController extends CoresecureController {
      */
     public function __construct(Request $request, ?array $space=null) {
         parent::__construct($request, $space);
-
         $this->spaceModel = new CoreSpace ();
     }