+++ tests/QuoteBaseTest.php
@@ -16,18 +16,19 @@ class QuoteBaseTest extends BaseTest {
     protected function activateQuote($space, $user) {
         Configuration::getLogger()->debug('activate quotes', ['user' => $user, 'space' => $space]);
         $this->asUser($user['login'], $space['id']);
-        $req = new Request([
+        $req = $this->request([
             "path" => "quoteconfig/".$space['id'],
             "formid" => "quotemenusactivationForm",
             "quoteMenustatus" => 3,
             "quoteDisplayMenu" => 0,
             "quoteDisplayColor" =>  "#000000",
             "quoteDisplayColorTxt" => "#ffffff"
-        ], false);
+        ]);
         $c = new QuoteconfigController($req, $space);
-        $c->indexAction($space['id']);
-        $c = new CorespaceController(new Request(["path" => "corespace/".$space['id']], false), $space);
-        $spaceView = $c->viewAction($space['id']);
+        $c->runAction('quote', 'index', ['id_space' => $space['id']]);
+        $req = $this->request(["path" => "corespace/".$space['id']]);
+        $c = new CorespaceController($req, $space);
+        $spaceView = $c->runAction('core', 'view', ['id_space' => $space['id']]);
         $quoteEnabled = false;
         foreach($spaceView['spaceMenuItems'] as $menu) {
             if($menu['url'] == 'quote') {
@@ -39,30 +40,30 @@ class QuoteBaseTest extends BaseTest {
     }
 
     protected function createQuoteUser($space, $user, $client):int {
-        $req = new Request([
+        $req = $this->request([
             "path" => "quoteuser/".$space['id'],
             "formid" => "editexistinguserForm",
             "id_space" => $space['id'],
             "id_user" => $user['id'],
             "id_client" => $client['id'],
             'date_open' => date('Y-m-d')
-        ], false);
+        ]);
         $c = new QuotelistController($req, $space);
-        $data = $c->editexistinguserAction($space['id'], 0);
+        $data = $c->runAction('quote', 'editexistinguser', ['id_space' => $space['id'], 'id' => 0]);
         $this->assertTrue($data['quote']['id'] > 0);
         return $data['quote']['id'];
     }
 
     protected function getQuote($space, $id):array {
-        $req = new Request([
+        $req = $this->request([
             "path" => "quoteuser/".$space['id'],
-        ], false);
+        ]);
         $c = new QuotelistController($req, $space);
-        return $c->editAction($space['id'], $id);
+        return $c->runAction('quote', 'edit', ['id_space' => $space['id'], 'id' => $id]);
     }
 
     protected function addQuoteItem($space, $quote, $item): int{
-        $req = new Request([
+        $req = $this->request([
             "path" => "quoteedititem/".$space['id'],
             "formid" => "createItemForm",
             "id" => 0,
@@ -70,9 +71,9 @@ class QuoteBaseTest extends BaseTest {
             "id_item" => $item['id'],
             "quantity" => $item['quantity'],
             "comment" => $item['comment']
-        ], false);
+        ]);
         $c = new QuotelistController($req, $space);
-        $data = $c->edititemAction($space['id']);
+        $data = $c->runAction('quote', 'edititem', ['id_space' => $space['id']]);
         $this->assertTrue($data['item']['id'] > 0);
         return $data['item']['id'];        
     }